+++ app/src/main/java/com/example/mindstone/ui/home/diary/DiaryEditFragment.kt
@@ -1,13 +1,14 @@
-package com.example.mindstone.ui.home
+package com.example.mindstone.ui.home.diary
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.core.view.ViewCompat
+import androidx.core.view.WindowInsetsCompat
 import androidx.fragment.app.activityViewModels
 import com.example.mindstone.databinding.FragmentDiaryEditBinding
-import com.example.mindstone.DiaryViewModel
 
 
 class DiaryEditFragment : Fragment() {
@@ -30,6 +31,14 @@ class DiaryEditFragment : Fragment() {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
+
+        // 화면 맞춤
+        ViewCompat.setOnApplyWindowInsetsListener(binding.root) { v, insets ->
+            val systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars())
+            v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom)
+            insets
+        }
+
         initClicker()
         changeVisibility()
         setObserver()
