+++ app/src/main/res/layout/fragment_diary_edit.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:background="#FFFFFF"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".ui.home.DiaryEditFragment">
+    tools:context=".ui.home.diary.DiaryEditFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
