+++ app/src/main/res/layout/fragment_diary_loading.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:background="#FFFFFF"
-    tools:context=".ui.home.DiaryLoadingFragment">
+    tools:context=".ui.home.diary.DiaryLoadingFragment">
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent">
